+++ app/src/main/java/com/example/hhhhhh/Fragment4.java
@@ -27,39 +27,12 @@ import android.widget.Spinner;
 import java.util.ArrayList;
 
 public class Fragment4 extends Fragment {
-    SellAdapter adatper;
-    ListView sellview;
-    ArrayList<SellItem> chatRoomList;
-    Button addbtn;
-    Spinner category_spinner;
+
 
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         setHasOptionsMenu(true);
-
         View v = inflater.inflate(R.layout.fragment_4, container, false);
-        chatRoomList = new ArrayList<SellItem>();
-        addbtn = (Button)v.findViewById(R.id.plusbtn);
-        sellview = (ListView) v.findViewById(R.id.lv4);
-        adatper = new SellAdapter();
-        sellview.setAdapter(adatper);
-        sellview.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-            @Override
-            public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
-                Intent intent = new Intent(getActivity(), Sellcontent.class);
-
-                startActivity(intent);
-            }
-        });
-        addbtn.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(getActivity(), PostingActivity.class);
-                startActivity(intent);
-            }
-        });
-        category_spinner = (Spinner)v.findViewById(R.id.fragment4_spinner);
-
 
 
         return v;
