+++ app/src/main/res/layout/fragment_4.xml
@@ -1,44 +1,14 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
-    android:orientation="vertical"
-    android:layout_height="match_parent">
+    android:layout_height="match_parent"
+    tools:context=".Fragment2">
 
-    <RelativeLayout
-        android:layout_width="match_parent"
-        android:layout_height="70dp">
-
-
-        <Spinner
-            android:id="@+id/fragment4_spinner"
-            android:layout_width="150dp"
-            android:layout_height="match_parent"
-            android:layout_alignParentLeft="true"
-            android:layout_marginLeft="8dp"
-            />
-
-        <Button
-            android:id="@+id/plusbtn"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_alignParentTop="true"
-
-            android:layout_alignParentRight="true"
-            android:layout_marginTop="12dp"
-            android:layout_marginRight="15dp"
-            android:text="추가하기" />
-
-
-    </RelativeLayout>
-
-    <View
-        android:layout_width="match_parent"
-        android:layout_height="2dp"
-        android:background="#9e9d9d"/>
-
-    <ListView
+    <!-- TODO: Update blank fragment layout -->
+    <TextView
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:id="@+id/lv4"/>
-</LinearLayout>
+        android:text="@string/hello_blank_fragment" />
+
+</FrameLayout>
